Locations can now be clicked and View controllers are notified
When clicking on a location o the AR View Controller, another View
Controller is displayed from the MainViewController. This simulates a
screen that could be used if a user clicks on it.   The delegate fires
and it's captured from the ARViewController, and then passed back to
whomever creates the ARViewController.  This should give greater
control. Also the ARGeoCoordinate is used in instead of the
ARCoordinate. Not sure why this was changed