Added AVFoundation to use for Camera viewing instead of UIImagePickerController
AVFoundation gives us greater control of the camera and will allow for
unlimited types of AR apps to be developed.
Gets rid of the Camera open animation
No longer loads as ModalViewController for the Camera view
Added another close button to show Peter how to close the new view and
the AR view.