Now updates distance of values within 2 meters
Added filters to only get updates of the Headings every 1 degree and to
update the location position every 2 meters.
Also the accuracy was set to within 10 meters.  This allows for the
less sampling of the data which was overkill and caused the app to draw
too many times a sec causing it to feel like it was shaking.
The distance is now recalculated roughly every 2 meters.  This cause be
within 10 meters of accuracy though.  Unless you need to sample within
10 meters this should be more than enough.