Added SubTitle that updates the distance when ever the location of the device changes
Added a SubTitle at the request of someone who wanted to know how.  The
ARView now has an updateView method that is called each time the
location of the device changes.
Cleaned up some of the code.
Formatted the distance displayed of the geolocation to be in kilometers
with 2 point decimal places.  Previously was showing in meters with
multiple decimal places.